requirements.txt
Django==2.1.7
django-markdownx==2.0.28
django-registration==3.0.1
Markdown==3.1
Pillow==6.0.0
urllib3==1.24.3
